requirements.txt
numpy
Pillow
torch
torchvision
matplotlib
tqdm
scikit-learn
wandb